proxy: Implement {Connect,Disconnect}Shim commands
These new commands are used by a shim to identify itself and signal the
proxy to clean up it's per-shim resources.

We do check that only a single process can claim a Token. More work can
be done to isolate a shim to a sub-set of commands but that will be the
object of a different commit.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>